+++ test/integration/test_clean_needed.py
@@ -19,7 +19,10 @@ class TestCleanNeeded(IntegrationTest):
         self.assertExists('mopack/logs/hello.log')
         self.assertExists('mopack/mopack.json')
 
-        self.assertPkgConfigLinkage('greeter')
+        self.assertPkgConfigLinkage('greeter', pkg_config_path=[
+            os.path.join('build', 'greeter', 'pkgconfig'),
+            os.path.join('build', 'hello', 'pkgconfig'),
+        ])
         self.assertPkgConfigLinkage('hello')
 
         output = json.loads(slurp('mopack/mopack.json'))
@@ -34,6 +37,7 @@ class TestCleanNeeded(IntegrationTest):
                 ),
                 cfg_directory_pkg(
                     'greeter', config,
+                    dependencies=['hello'],
                     path={'base': 'cfgdir', 'path': 'greeter-bfg'},
                     builders=[cfg_bfg9000_builder(extra_args=['--extra'])],
                     linkage=cfg_pkg_config_linkage(pcname='greeter')
@@ -53,7 +57,10 @@ class TestCleanNeeded(IntegrationTest):
         self.assertNotExists('mopack/build/greeter/extra.txt')
         self.assertNotExists('mopack/build/hello/extra.txt')
 
-        self.assertPkgConfigLinkage('greeter')
+        self.assertPkgConfigLinkage('greeter', pkg_config_path=[
+            os.path.join('build', 'greeter', 'pkgconfig'),
+            os.path.join('build', 'hello', 'pkgconfig'),
+        ])
         self.assertPkgConfigLinkage('hello')
 
         output = json.loads(slurp('mopack/mopack.json'))
@@ -72,6 +79,7 @@ class TestCleanNeeded(IntegrationTest):
                 ),
                 cfg_directory_pkg(
                     'greeter', config,
+                    dependencies=['hello'],
                     path={'base': 'cfgdir', 'path': 'greeter-bfg'},
                     builders=[cfg_bfg9000_builder()],
                     linkage=cfg_pkg_config_linkage(pcname='greeter')