+++ test/unit/linkages/test_path_system.py
@@ -59,12 +59,11 @@ class TestPath(LinkageTest):
         if os.path.exists(self.pkgdir):
             shutil.rmtree(self.pkgdir)
 
-    def check_linkage(self, linkage, *, name='foo', dependencies=[],
-                      include_path=[], library_path=[], headers=[],
-                      libraries=None, compile_flags=[], link_flags=[]):
+    def check_linkage(self, linkage, *, name='foo', include_path=[],
+                      library_path=[], headers=[], libraries=None,
+                      compile_flags=[], link_flags=[]):
         if libraries is None:
             libraries = [name]
-        self.assertEqual(linkage.dependencies, dependencies)
         self.assertEqual(linkage.include_path, include_path)
         self.assertEqual(linkage.library_path, library_path)
         self.assertEqual(linkage.headers, headers)
@@ -226,22 +225,27 @@ class TestPath(LinkageTest):
             self.make_linkage('foo', version={'type': 'goofy'})
 
     def test_dependencies(self):
-        linkage = self.make_linkage('foo', dependencies=['bar'])
-        self.check_linkage(linkage, dependencies=[Dependency('bar')])
-        self.check_get_linkage(linkage, 'foo', None)
+        pkg = MockPackage('foo', srcdir=self.srcdir, builddir=self.builddir,
+                          dependencies=[Dependency('bar')],
+                          _options=self.make_options())
+
+        linkage = self.make_linkage(pkg)
+        self.check_linkage(linkage)
+        self.check_get_linkage(linkage, 'foo', None, pkg=pkg)
         self.check_pkg_config('foo', None, {
             'libs': ['-L' + abspath('/mock/lib'), '-lfoo', '-lbar'],
         })
 
+        # Try getting linkage again with a different dependency linkage.
         path = '/mock/pkgconfig'
-        dep_linkage = {'name': 'foo', 'type': self.type, 'pcnames': ['bar'],
+        dep_linkage = {'name': 'bar', 'type': self.type, 'pcnames': ['bar'],
                        'pkg_config_path': [path]}
         with mock.patch('mopack.origins.Package.get_linkage',
                         return_value=dep_linkage):
             self.check_get_linkage(linkage, 'foo', None, {
                 'name': 'foo', 'type': self.type, 'pcnames': ['foo'],
                 'pkg_config_path': [self.pkgconfdir, path],
-            })
+            }, pkg=pkg)
 
     def test_include_path_relative(self):
         pkg = MockPackage(srcdir=self.srcdir, builddir=self.builddir)