+++ test/unit/origins/sdist/test_directory.py
@@ -16,6 +16,7 @@ from mopack.linkages.path_system import SystemLinkage
 from mopack.origins import Package
 from mopack.origins.apt import AptPackage
 from mopack.origins.sdist import DirectoryPackage
+from mopack.origins.submodules import UnmanagedSubmoduleProps
 from mopack.path import Path
 from mopack.types import ConfigurationError, FieldError, Unset
 from mopack.yaml_tools import SafeLineLoader, YamlParseError
@@ -626,7 +627,9 @@ class TestDirectory(SDistTestCase):
             self.assertIsInstance(pkg, DirectoryPackage)
             self.assertIsInstance(pkg.linkage, SystemLinkage)
             self.assertEqual([type(i) for i in pkg.builders], [NoneBuilder])
-            self.assertEqual(pkg.submodules, {'sub': {}})
+            self.assertEqual(pkg.submodules, {
+                'sub': UnmanagedSubmoduleProps(opts.expr_symbols)
+            })
             self.assertEqual(pkg.submodule_required, False)
             m.assert_called_once()