+++ test/unit/origins/sdist/test_git.py
@@ -15,6 +15,7 @@ from mopack.linkages.path_system import SystemLinkage
 from mopack.origins import Package
 from mopack.origins.apt import AptPackage
 from mopack.origins.sdist import GitPackage
+from mopack.origins.submodules import UnmanagedSubmoduleProps
 from mopack.types import ConfigurationError, Unset
 
 
@@ -693,7 +694,9 @@ class TestGit(SDistTestCase):
             self.assertIsInstance(pkg, GitPackage)
             self.assertIsInstance(pkg.linkage, SystemLinkage)
             self.assertEqual([type(i) for i in pkg.builders], [NoneBuilder])
-            self.assertEqual(pkg.submodules, {'sub': {}})
+            self.assertEqual(pkg.submodules, {
+                'sub': UnmanagedSubmoduleProps(opts.expr_symbols)
+            })
             self.assertEqual(pkg.submodule_required, False)
             m.assert_called_once()