+++ test/unit/origins/test_conan.py
@@ -12,6 +12,7 @@ from mopack.iterutils import iterate
 from mopack.origins import Package, PackageOptions
 from mopack.origins.apt import AptPackage
 from mopack.origins.conan import ConanPackage
+from mopack.origins.submodules import ManagedSubmoduleProps
 from mopack.shell import ShellArguments
 
 
@@ -416,7 +417,9 @@ class TestConan(OriginTest):
                                side_effect=ConanPackage.upgrade) as m:
             pkg = Package.rehydrate(data, _options=opts, **rehydrate_kwargs)
             self.assertIsInstance(pkg, ConanPackage)
-            self.assertEqual(pkg.submodules, {'sub': {}})
+            self.assertEqual(pkg.submodules, {
+                'sub': ManagedSubmoduleProps(opts.expr_symbols)
+            })
             self.assertEqual(pkg.submodule_required, False)
             m.assert_called_once()